% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ll.R
\name{fit_ll}
\alias{fit_ll}
\title{Fit local linear estimator.}
\usage{
fit_ll(x, y, h, kernel)
}
\arguments{
\item{x}{\code{numeric(n)} or \code{matrix(, n, p)}. n observations of
p-dimensional predictors.}

\item{y}{\code{numeric(n)} or \code{matrix(, n, k)}. n observations of
k-dimensional predictors.}

\item{h}{\code{numeric(1)}. The bandwidth to use.}

\item{kernel}{\code{function}. The kernel.}
}
\value{
A numeric n by k matrix. The estimates of response variable at the
locations of the observed predictors.
}
\description{
Take predictor-response pairs and estimate new responses for the same
predictors via a local linear estimator. Both, predictors and responses, may
be multidimensional.
}
