% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLocal.R
\name{fitLocalLinear}
\alias{fitLocalLinear}
\title{Fit local linear estimator.}
\usage{
fitLocalLinear(x, y, bw, kernel)
}
\arguments{
\item{x}{\code{numeric(n)} or \code{matrix(, n, p)}. n observations of
p-dimensional predictors.}

\item{y}{\code{numeric(n)} or \code{matrix(, n, k)}. n observations of
k-dimensional responses.}

\item{bw}{\code{numeric(1)}. The bandwidth to use.}

\item{kernel}{\code{function}. The kernel. Maps the Euclidean distance of
points in the predictor space to a weight.}
}
\value{
A numeric n by k matrix. The estimates of response variable at the
locations of the observed predictors.
}
\description{
Take predictor-response pairs and estimate new responses for the same
predictors via a local linear estimator. Both, predictors and responses, may
be multidimensional.
}
